# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Rooms Booking Service",
        "contact": {}
    },
    "host": "127.0.0.1:8080",
    "paths": {
...
        "/rooms/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Список переговорок (admin и user)",
                "responses": {
                    "200": {
                        "description": "Список переговорок",
                        "schema": {
                            "$ref": "#/definitions/internal_features_rooms_transport_http.RoomsDTOResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.InternalError"
                        }
                    }
                }
            }
        },
        "/rooms/{roomId}/schedule/create": {
            "post": {
                "description": "Доступно только роли admin. Расписание можно создать только один раз для каждой переговорки. Поле daysOfWeek должно содержать значения от 1 до 7 (1=Пн, 7=Вс). При передаче значений вне этого диапазона возвращается 400.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Создать расписание переговорки (только admin, только один раз). Длительность слота 30 мин. После создания расписание изменить нельзя.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор переговорки",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Расписание сохранено",
                        "schema": {
                            "$ref": "#/definitions/internal_features_schedules_transport_http.ScheduleDTOResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос (в т.ч. недопустимые значения daysOfWeek)",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён (требуется роль admin)",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "404": {
                        "description": "Переговорка не найдена",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "409": {
                        "description": "Расписание для переговорки уже создано, изменение не допускается",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.InternalError"
                        }
                    }
                }
            }
        },
        "/rooms/{roomId}/slots/list": {
            "get": {
                "description": "Возвращает слоты, не занятые активной бронью, для указанной переговорки на указанную дату. Все даты и время передаются и возвращаются в UTC. Параметр date является обязательным; при его отсутствии возвращается 400. Если у переговорки нет расписания — возвращается пустой список (переговорка считается всегда недоступной).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slots"
                ],
                "summary": "Список доступных для бронирования слотов по переговорке и дате (admin и user). Наиболее нагруженный эндпоинт.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор переговорки",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата в формате ISO 8601 (например: 2024-06-10). Обязательный параметр; при отсутствии — 400",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список доступных слотов (не занятых активной бронью). Пустой список, если у переговорки нет расписания или на эту дату нет слотов.",
                        "schema": {
                            "$ref": "#/definitions/internal_features_slots_transport_http.SlotsDTOResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос (отсутствует или некорректен параметр date)",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "404": {
                        "description": "Переговорка не найдена",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.InternalError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "backend-assignment-avito_internal_core_domain.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "enum": [
                        "INVALID_REQUEST",
                        " INTERNAL_ERROR",
                        " SCHEDULE_EXISTS",
                        " SLOT_ALREADY_BOOKED",
                        " FORBIDEN",
                        " NOT_FOUND"
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
        "backend-assignment-avito_internal_core_domain.InternalError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "enum": [
                        "INTERNAL_ERROR"
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
        "internal_features_rooms_transport_http.RoomDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "0",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "name": {
                    "type": "string",
                    "x-order": "1",
                    "example": "work_room"
                },
                "description": {
                    "type": "string",
                    "x-order": "2",
                    "example": "room for work..."
                },
                "capacity": {
                    "type": "integer",
                    "x-order": "3",
                    "example": 3
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "4",
                    "example": "2026-03-25T12:00:41.267Z"
                }
            }
        },
...
        "internal_features_rooms_transport_http.RoomsDTOResponse": {
            "type": "object",
            "properties": {
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_features_rooms_transport_http.RoomDTO"
                    }
                }
            }
        },
        "internal_features_schedules_transport_http.ScheduleDTOResponse": {
            "type": "object",
            "properties": {
                "daysOfWeek": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                }
            }
        },
        "internal_features_slots_transport_http.SlotDTOResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "0",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "roomId": {
                    "type": "string",
                    "x-order": "1",
                    "example": "78a85f64-6717-4562-b3fc-2c963f6abfa6"
                },
                "start": {
                    "type": "string",
                    "x-order": "2",
                    "example": "2026-03-25T14:01:22.680Z"
                },
                "end": {
                    "type": "string",
                    "x-order": "3",
                    "example": "2026-03-25T14:01:22.680Z"
                }
            }
        },
        "internal_features_slots_transport_http.SlotsDTOResponse": {
            "type": "object",
            "properties": {
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_features_slots_transport_http.SlotDTOResponse"
                    }
                }
            }
        }
    }
}